+++ pkg/cloudentity/client.go
