+++ pkg/envoy/config/listener.go
@@ -37,7 +37,7 @@ import (
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 	"google.golang.org/protobuf/types/known/anypb"
 
-	"github.com/kubeshop/kusk-gateway/internal/cert"
+	"github.com/kubeshop/kusk-gateway/pkg/cert"
 )
 
 const (
