+++ pkg/envoy/manager/envoy_config_manager.go
