+++ pkg/envoy/types/cors.go
