+++ pkg/envoy/types/cors_test.go
