+++ pkg/mocking/text.go
